+++ packages/react-router-manage/src/components/NoAuth/index.tsx
@@ -2,16 +2,16 @@ import * as React from 'react';
 
 const PageConfig = {
   '401': {
-    title: '无权限查看该页面',
+    title: '401 NO PERMISSION',
     img: '//ysf.qiyukf.net/yx/9c9ce7793b3c0657da5d80e740a89681',
   },
   '404': {
-    title: '页面不存在',
+    title: '404 NOT FOUND',
     img: 'https://ysf.nosdn.127.net/ysh/6be90dea7806767fe65e7b48982b7a61',
   },
 } as Record<string, {title: string; img: string}>;
 
-const NoAuth: React.FC<{ code?: string }> = ({ code = '401' }) => {
+const NoAuth: React.FC<{ code?: '404' | '401' }> = ({ code = '401' }) => {
   const config = PageConfig[code];
   return (
     <div